refactor: split ChannelMasthead.vue
Closes #989. Fourth child of #980.

The masthead spelled every variant out inline: the title zone with its three
avatar treatments, the connection pill, the roster facepile and the options
menu all shared one scope. They move out along the seams that were already
there.

- components/channel/MastheadTitle.vue: the avatar, the name, the
  notification cue and the sub-lines under them
- components/channel/MastheadOptionsMenu.vue: the kebab menu's three sections
- components/channel/MastheadFacepile.vue: the roster stack and its readout
- components/channel/MastheadConnectionPill.vue: the two socket cues
- composables/useMastheadSearch.ts: where the search glyph sends the viewer

MAX_MASTHEAD_AVATARS moves to lib/memberAvatars and the active-member count
to lib/presence, so the two stacks and the two readouts cannot disagree.

Its entry is gone from the max-lines exemption block in eslint.config.js.
