+++ resources/js/lib/presence.ts
@@ -23,6 +23,22 @@ export function dmParticipantPresence(
 }
 
 /**
+ * How many of a roster are active right now.
+ *
+ * Away counts as present but not active, which is the whole point of the third
+ * state: the readout answers "who could answer me now", not "who has the tab
+ * open". Shared by the masthead's facepile and the compact readout that stands
+ * in for it below the breakpoint, which must never disagree.
+ */
+export function activeMemberCount(
+    members: Array<{ id: string }>,
+    presenceFor: (userId: string) => RenderedPresence,
+): number {
+    return members.filter((member) => presenceFor(member.id) === 'active')
+        .length;
+}
+
+/**
  * The message key announcing a presence to assistive tech.
  *
  * Returned untranslated so the caller resolves it through `$t` and the label
